+++ api/src/certification/evaluation/routes.js
@@ -1,7 +1,7 @@
 import * as companionAlert from '../evaluation/application/companion-alert-route.js';
+import * as certificationAdminRoutes from './application/certification-admin-route.js';
 import * as certificationCourseRoutes from './application/certification-course-route.js';
-import * as certificationRoutes from './application/certification-route.js';
 
-const certificationEvaluationRoutes = [companionAlert, certificationCourseRoutes, certificationRoutes];
+const certificationEvaluationRoutes = [companionAlert, certificationCourseRoutes, certificationAdminRoutes];
 
 export { certificationEvaluationRoutes };
